EAppUpdateStatus.updateIncomplete
